[heap] Use SafeHeapObjectSize in more places
In particular use it for iterating live object ranges and rework the
affected callsites.

Bug: 425896305
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6719831
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101342}

+++ src/heap/scavenger-inl.h
@@ -66,8 +66,7 @@ bool Scavenger::MigrateObject(Tagged<Map> map, Tagged<HeapObject> source,
   // memory ordering for the CAS above.
   target->set_map_word(map, kRelaxedStore);
   heap()->CopyBlock(target.address() + kTaggedSize,
-                    source.address() + kTaggedSize,
-                    static_cast<size_t>(size.value() - kTaggedSize));
+                    source.address() + kTaggedSize, size.value() - kTaggedSize);
 
   if (V8_UNLIKELY(is_logging_)) {
     // TODO(425150995): We should have uint versions for allocation to avoid
